Initial support for terminal commandline console.
- Added CommandlinePlugin to lib.rs.
- Added file src/commandline.rs to handle commandline functionality.

Commandline is a way to use bevy-console in headless environment (no gui available) for example on a server or possibly in the future as lighter alternative for the egui version.

The current implementation of terminal commandline console is built on top of the existing code and doesn't modify how the console (egui) version of bevy-console works. Commandline is meant to be an additional feature that should not be required for any other part of bevy-console.

Ideally the commandline should work without the egui stuff, but that would probably require some refactoring of the code.